cambios David y añadir buscador en search
