+++ app/src/main/java/com/example/myfavbook/activities/Search.java
@@ -1,4 +1,4 @@
-package com.example.myfavbook;
+package com.example.myfavbook.activities;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -7,36 +7,38 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
 
+import com.example.myfavbook.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
-public class MainHome extends AppCompatActivity {
+public class Search extends AppCompatActivity {
 
     BottomNavigationView bottomNavigationView;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main_home);
-        bottomNavigationView = findViewById(R.id.bottom_navigator);
-        bottomNavigationView.setSelectedItemId(R.id.home);
+        setContentView(R.layout.activity_search);
+
+        bottomNavigationView = findViewById((R.id.btn_navigator));
+        bottomNavigationView.setSelectedItemId(R.id.search);
 
         bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem item) {
                 switch (item.getItemId())
                 {
-                    case R.id.home:
+                    case R.id.search:
 
+                        return true;
+                    case R.id.home:
+                        startActivity(new Intent(getApplicationContext(),MainHome.class));
+                        overridePendingTransition(0,0);
                         return true;
 
                     case R.id.more:
                         startActivity(new Intent(getApplicationContext(),More.class));
                         overridePendingTransition(0,0);
                         return true;
-                    case R.id.search:
-                        startActivity(new Intent(getApplicationContext(),Search.class));
-                        overridePendingTransition(0,0);
-                        return true;
+
                     case R.id.my_books:
                         startActivity(new Intent(getApplicationContext(),Books.class));
                         overridePendingTransition(0,0);
@@ -45,5 +47,10 @@ public class MainHome extends AppCompatActivity {
                 return false;
             }
         });
+
+
+
+
+
     }
 }
