+++ app/src/main/res/layout/activity_books.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Books">
+    android:background="@drawable/background_new"
+    tools:context=".activities.Books">
 
     <TextView
         android:layout_width="wrap_content"
@@ -18,7 +19,7 @@
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:id="@+id/bottom_navigator"
+        android:id="@+id/btn_navigator"
         app:itemBackground="@color/blue_claro"
         app:itemTextColor="@drawable/selector"
         app:itemIconTint="@drawable/selector"
