+++ app/src/main/res/layout/activity_main.xml
@@ -20,13 +20,13 @@
         app:layout_constraintBottom_toTopOf="@+id/btnAcceder"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintTop_toTopOf="@id/imageView"
         app:layout_constraintVertical_bias="0.812" />
 
     <Button
         android:id="@+id/btnAcceder"
-        android:layout_width="211dp"
-        android:layout_height="wrap_content"
+        android:layout_width="315dp"
+        android:layout_height="89dp"
         android:layout_marginStart="135dp"
         android:layout_marginTop="350dp"
         android:layout_marginEnd="135dp"
@@ -34,12 +34,24 @@
         android:background="#33339A"
         android:onClick="goToHome"
         android:text="@string/acceder"
-        android:textSize="16sp"
+        android:textSize="28sp"
         android:visibility="visible"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.503"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.078" />
+        app:layout_constraintVertical_bias="0.57" />
+
+    <ImageView
+        android:id="@+id/imageView"
+        android:layout_width="wrap_content"
+        android:layout_height="0dp"
+        android:layout_marginBottom="24dp"
+        app:layout_constraintBottom_toTopOf="@+id/textView4"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.51"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.0"
+        app:srcCompat="@drawable/frase" />
 </androidx.constraintlayout.widget.ConstraintLayout>
