+++ app/src/main/res/layout/activity_main_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainHome"
+    tools:context=".activities.MainHome"
     android:background="@drawable/background_new">
 
     <TextView
@@ -19,7 +19,7 @@
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:id="@+id/bottom_navigator"
+        android:id="@+id/btn_navigator"
         app:itemBackground="@color/blue_claro"
         app:itemTextColor="@drawable/selector"
         app:itemIconTint="@drawable/selector"
